adding function update package for admin

+++ src/main/java/com/restaurent/manager/repository/PackageRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface PackageRepository extends JpaRepository<Package,Integer> {
+public interface PackageRepository extends JpaRepository<Package,Long> {
     Package findByPackName(String packName);
 }
